Revert "710: optmize memory: use int32"
This reverts commit 5879042711b11f8e83e4a3a26c9a70eeacc826c9.

+++ leetcode/710/solver/solver.go
@@ -4,28 +4,28 @@ import "math/rand"
 
 type Solution struct {
 	n                int32
-	blockedToAllowed map[int32]int32
+	blockedToAllowed map[int]int
 }
 
 func Constructor(n int, blacklist []int) Solution {
 	divider := n - len(blacklist)
-	tempAllowList := make([]int32, 0, n-divider)
+	tempAllowList := make([]int, 0, n-divider)
 	for i := divider; i < n; i++ {
-		tempAllowList = append(tempAllowList, int32(i))
+		tempAllowList = append(tempAllowList, i)
 	}
 
 	for _, v := range blacklist {
 		if v >= divider {
-			tempAllowList = remove(tempAllowList, find(tempAllowList, int32(v)))
+			tempAllowList = remove(tempAllowList, find(tempAllowList, v))
 		}
 	}
 
-	blockedToAllowed := make(map[int32]int32)
-	iterator := int32(0)
+	blockedToAllowed := make(map[int]int)
+	iterator := 0
 
 	for _, v := range blacklist {
 		if v < divider {
-			blockedToAllowed[int32(v)] = tempAllowList[iterator]
+			blockedToAllowed[v] = tempAllowList[iterator]
 			iterator++
 		}
 	}
@@ -35,26 +35,26 @@ func Constructor(n int, blacklist []int) Solution {
 
 func (this *Solution) Pick() int {
 	num := rand.Int31n(this.n)
-	replacement, found := this.blockedToAllowed[num]
+	replacement, found := this.blockedToAllowed[int(num)]
 
 	if found {
-		return int(replacement)
+		return replacement
 	}
 
 	return int(num)
 }
 
-func find(s []int32, v int32) int32 {
+func find(s []int, v int) int {
 	for i := range s {
 		if s[i] == v {
-			return int32(i)
+			return i
 		}
 	}
 
 	return -1
 }
 
-func remove(s []int32, i int32) []int32 {
+func remove(s []int, i int) []int {
 	s[i] = s[len(s)-1]
 	return s[:len(s)-1]
 }
